# server/docs/swagger.yaml
basePath: /
definitions:
...
  types.PaymentStatus:
    enum:
    - pending
    - proof_submitted
    - creditor_confirmed
    - disputed
    - settled
    type: string
    x-enum-varnames:
    - PaymentPending
    - PaymentProofSubmitted
    - PaymentCreditorConfirmed
    - PaymentDisputed
    - PaymentSettled
  types.PaymentView:
    properties:
      amount_baisa:
        type: integer
      created_at:
        type: string
      from:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/types.PaymentStatus'
      to:
        type: string
    type: object
...
  types.ProofType:
    enum:
    - image
    - text
    type: string
    x-enum-varnames:
    - ProofImage
    - ProofText
  types.ProofView:
    properties:
      byte_size:
        type: integer
      created_at:
        type: string
      note:
        type: string
      payment_id:
        type: string
      proof_type:
        $ref: '#/definitions/types.ProofType'
      sha256:
        type: string
    type: object
...
  types.SubmitProofRequest:
    properties:
      note:
        type: string
    type: object
...
  types.apiError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Shared-expense settlement tracker for group trips. Members record what they paid; closing a group computes each member's fair share and the minimum set of payments that squares everyone up. Money moves off-platform — each computed payment is tracked through a two-key confirmation flow (debtor proves, creditor attests, admin finalizes).

    All amounts are integer **baisa** (1.000 OMR = 1000). There are no floats in the money path.

    **Authenticating here:** call `POST /auth/login` (try `admin@expense-splitter.local` / `admin`), copy the `access_token` from the response, click **Authorize** above and paste `Bearer <access_token>`.
  title: Expense Splitter API
  version: "1.0"
paths:
...
  /payments/{id}/confirm:
    post:
      description: '**Key 2 of 2.** The creditor — the person owed — confirms the
        money actually arrived, moving `proof_submitted` -> `creditor_confirmed`.
        This is the *only* transition that produces `creditor_confirmed`, and only
        the creditor may call it.'
      parameters:
      - description: payment id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: payment now creditor_confirmed
          schema:
            $ref: '#/definitions/types.PaymentView'
        "400":
          description: invalid payment id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: you are not this payment's creditor
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: payment not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: payment is not awaiting your confirmation
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Attest that you received the money (creditor)
      tags:
      - payments
  /payments/{id}/dispute:
    post:
      description: The creditor says the money never arrived, moving the payment to
        `disputed`. A disputed payment can **never** be reported settled — the debtor
        must re-submit proof and both keys must run again.
      parameters:
      - description: payment id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: payment now disputed
          schema:
            $ref: '#/definitions/types.PaymentView'
        "400":
          description: invalid payment id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: you are not this payment's creditor
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: payment not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: payment is not in a disputable state
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Deny receiving the money (creditor)
      tags:
      - payments
  /payments/{id}/finalize:
    post:
      description: |-
        The group-admin (own group) or global admin (any group) marks a `creditor_confirmed` payment as **settled** — the only transition that produces `settled`, and it is terminal.

        A debtor can never settle their own payment, **even if they are also the group-admin**. The global admin's override may finalize any non-settled payment and can do nothing else.
      parameters:
      - description: payment id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: payment now settled
          schema:
            $ref: '#/definitions/types.PaymentView'
        "400":
          description: invalid payment id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not an admin for this payment's group, or you are the debtor
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: payment not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: payment has not been confirmed by the creditor
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Finalize a payment as settled (admin)
      tags:
      - payments
  /payments/{id}/proof:
    get:
      description: Visible to the debtor, the creditor, and admins. Returns the note
        and/or the image's metadata — content type, size, and sha256. Use `/proof/image`
        for the bytes.
      parameters:
      - description: payment id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: proof note and/or image metadata
          schema:
            $ref: '#/definitions/types.ProofView'
        "400":
          description: invalid payment id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not a party to this payment
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: payment or proof not found
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Get a payment's proof metadata
      tags:
      - payments
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: |-
        **Key 1 of 2.** The debtor evidences that they paid their creditor in real life, moving the payment to `proof_submitted`. Also the way out of `disputed` — re-submit and both keys run again.

        Accepts either shape:
        - **JSON** `{"note": "paid 40 cash on day 3"}`
        - **multipart/form-data** with an `image` file field (receipt or transfer screenshot)

        Images are capped at 5 MiB and validated by **magic bytes** (jpeg/png/gif/webp) — never by file extension. The stored sha256 makes a later swap detectable.
      parameters:
      - description: payment id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: text note (JSON variant)
        in: body
        name: body
        schema:
          $ref: '#/definitions/types.SubmitProofRequest'
      - description: receipt image, max 5 MiB (multipart variant)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: payment now proof_submitted
          schema:
            $ref: '#/definitions/types.PaymentView'
        "400":
          description: invalid id/body, oversized file, or not a real image
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: you are not this payment's debtor
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: payment not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: payment is not in a state that accepts proof
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Upload proof of payment (debtor)
      tags:
      - payments
  /payments/{id}/proof/image:
    get:
      description: Visible to the debtor, the creditor, and admins. Streams the stored
        image from object storage with its real content type.
      parameters:
      - description: payment id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/jpeg
      - image/png
      responses:
        "200":
          description: the proof image
          schema:
            type: file
        "400":
          description: invalid payment id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not a party to this payment
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: no image proof on this payment
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Download the proof image bytes
      tags:
      - payments
  /payments/{id}/reject:
    post:
      description: The group-admin (own group) or global admin (any group) sends a
        payment to `disputed`, returning it to the debtor to re-submit proof. `settled`
        stays terminal — even for the global admin.
      parameters:
      - description: payment id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: payment now disputed
          schema:
            $ref: '#/definitions/types.PaymentView'
        "400":
          description: invalid payment id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not an admin for this payment's group
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: payment not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: payment is already settled
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Reject a payment back to disputed (admin)
      tags:
      - payments
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Bearer token from POST /auth/login. Format: `Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...